refactor(host-agent): move server-direction desktop wire types out of shed-core
The slice-1a server-direction wire types (hello_ack/approval_request/event/ping/
token.response builders + decode_desktop + their structs) were added to
shed-core::approval::protocol mirroring the client direction — but only the host-agent
daemon's desktop server uses them, and the desktop apps link shed-core via FFI without
ever calling them (~0.2 MB of dead-weight in the app). Move them to a new
crates/shed-host-agent/src/desktop_protocol.rs (feature-gated behind desktop-forwarding,
so the headless build drops them too).

- shed-core keeps only the CLIENT-direction wire types (hello/approval_response/pong/
  token.get builders + decode()→HostAgentInbound + HelloAck/ApprovalRequest/
  AuditEventFrame/TokenResponse), which shed-app's host_agent.rs client + the desktop
  apps genuinely share — the plan's don't-over-share boundary rule.
- Pure move: every emitted frame is byte-identical (the differential + golden pin them,
  incl. the superseded hello_ack empty-agent path); HOST_AGENT_PROTOCOL_VERSION stays
  the single source of truth in shed-core.
- Zeroes this port's desktop-app-size impact (61.3→61.1 MB, the pre-existing baseline).

Verified: shed-app/shed-core-ffi/shedctl reference no moved item; the FFI-exported
surface is unchanged.

Gate: cargo test --workspace (shed-core 131, shed-host-agent 73, golden 10, shed-app 65,
shedctl 13) + clippy clean; headless compiles; make test-host-agent-diff (16 + golden);
Go untouched.
